requirements.txt
fasttext-wheel
pandas==2.0.1
numpy==1.24.3
gunicorn==20.1.0
flask==2.3.2
scikit-learn==1.3.2
joblib==1.3.2
nltk==3.8.1
xlrd==2.0.1
openpyxl==3.1.2
boto3
hyperopt
seaborn
matplotlib
Pillow
beautifulsoup4
# fasttext==0.9.2